requirements.txt
Django==5.0.4
djangorestframework==3.15.1
mysql==0.0.3
mysql-connector-python==8.3.0
mysql-connector==2.2.9
mysqlclient==2.2.4
mysql-client==0.0.1